Task #65: Controller class for getting list
