+++ src/main/java/org/example/JavaTaigaCode/controllers/ProjectController.java
@@ -1,0 +1,29 @@
+package org.example.JavaTaigaCode.controllers;
+
+import org.example.JavaTaigaCode.models.ProjectDTO;
+import org.example.JavaTaigaCode.service.ProjectService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api")
+public class ProjectController {
+    @Autowired
+    ProjectService projectService;
+    @GetMapping("/getProjectList/{memberID}")
+    @ResponseBody
+    public List<ProjectDTO> getProjectList(@PathVariable("memberID") Integer memberID) {
+        List<ProjectDTO> projects = projectService.getPojectList(memberID);
+
+            if (projects != null) {
+                for (ProjectDTO project : projects) {
+                    System.out.println(project.toString());
+                }
+            } else {
+                throw new RuntimeException("Unable to get Project List for user");
+            }
+            return projects;
+    }
+}
